# Optimize Topological Sort for interfaces