The benchmark for sorting `.GetInterfaces()` over all model types
```
|             Method |         Mean | Error |          Min |          Max |    Gen 0 |  Allocated |
|------------------- |-------------:|------:|-------------:|-------------:|---------:|-----------:|
| OldTopologicalSort |    18.342 ms |    NA |    18.342 ms |    18.342 ms |        - |   6,308 KB |
| NewTopologicalSort |     3.939 ms |    NA |     3.939 ms |     3.939 ms |        - |     644 KB |
```

Also renamed `loops` param to `cycles`, because in Graph theory  "loop" means zero-length cycle

Existing Topological Sort algorithms remain in codebase because they are used in places, which cannot be easy converted to use the new algorithm with callback predicate and returning removed edges.