@botinko 2022-09-12
Just for information, on which operations this new optimized version of sort is used?